// src/types/RiskCalculation.ts
export enum Quality {
  CONSENSUS = "consensus",
  AVERAGE = "average",
  MISSING = "missing",
}